Added some output for the user when os.name is not correct
Added some std output for the user when os.name is not correct before the UnsupportedPlatformException is thrown.

"TESTAR detected OS: " + osName + " and this is not yet supported. If the detected OS is wrong, please contact the TESTAR team at info@testar.org. Exiting with Exception."

Not correct here means it does not contain the substring "Windows" or "Linux". After a quick google it seems that for some versions of Windows Server does not always return the right os.name string.