Fixed creating screenshots for the webdriver protocol.
Since getState is called before begineSecquence it should be save to
move the logic of setting the system PID and window handle towards
getState. This is needed because in getState we create an initial
screenshot and therefor we need to have the window handle in order to
obtain the display scale.