Fix a fascinating dialog bug
The SpyModePanel is using the TreeSetListModel class which overrides the "getElementAt(index)"method.

If the SpyModePanel is loaded before the model or llm panels (not sure which one exactly) a concurrent exception modification is thrown in the override "getElementAt(index)"method. Probably because some GUI element from the model or llm panel is also using this super.getElementAt(index) function

TreeSetListModel should be implemented thread-safe. In fact, this is more like a temp solution.